==> long-term original/Process_steps/step1_preprocess.ipynb <==
import argparse
import os
import sys
import psutil
import numpy as np
import pandas as pd
import matplotlib
matplotlib.use('Agg')
import matplotlib.pyplot as plt
#from datetime import datetime
import math
from IPython.display import display
from multiprocessing import cpu_count,Pool 
import multiprocessing
from joblib import Parallel, delayed
from orderedset import OrderedSet
import datetime

def print_memory_usage():
    print ("memory log:")
    process = psutil.Process(os.getpid())
    print("%5.2f GB (RSS)" % (process.memory_info().rss / 2**30))
    print("%5.2f GB (VMS)" % (process.memory_info().vms / 2**30))
    print("%5.2f GB (Used)" % (psutil.virtual_memory().used / 2**30))
    print("%5.2f GB (Available)" % (psutil.virtual_memory().available / 2**30))
    print("%5.2f GB (Total)" % (psutil.virtual_memory().total / 2**30))


def days_between(data):
    #print (multiprocessing.current_process())
    data['duration'] = abs((data.EndTime - data.StartTime).astype('timedelta64[s]')/60)
    return data

def type_modif(data):
    #created = multiprocessing.Process()
    #print ('created:', created.name, created._identity)
    #print ("Process created")
    data['RefinedType'] = data.RefinedType_.apply(lambda x: x.split('-')[0])
    return data

def State_set(data,map):
    #mask = (data.Type == 'W')
    #z_valid = data[mask]
    data['City']=data.apply(lambda row: map[row['AirportCode']]['City'] if (row['Type']=='W' and row['AirportCode'] in map.keys()) else row['City'],axis=1)
    data['State']=data.apply(lambda row: map[row['AirportCode']]['State'] if (row['Type']=='W' and row['AirportCode'] in map.keys()) else row['State'],axis=1)
    return data

class WithExtraArgs(object):
    def __init__(self, func, **args):
        self.func = func
        self.args = args
    def __call__(self, df):
        return self.func(df, **self.args)

def applyParallel(pool,data, func, kwargs):
    data_split = np.array_split(data, min(partitions,data.shape[0]))
    data = pd.concat(pool.map(WithExtraArgs(func, **kwargs), data_split))
    return data

def parallelize(data, func,pool):
    data_split = np.array_split(data, partitions)
    data = pd.concat(pool.map(func, data_split))
    return data


filepath = '../../AllEvents_Aug2016-Aug2018.csv'
raduis = 14 # mi

cores = cpu_count() #Number of CPU cores on your system
partitions = cores #Define as many partitions as you want
print ("number of cores "+str(partitions))

#load data
use_cols=['EventId','Type',	'RefinedType_',	'StartTime'	
    ,'EndTime',	'LocationLat',	'LocationLng',	'Distance(mi)'	
    ,'AirportCode',	'Number',	'Street',	'Side',	'City',	'County'	
    ,'State',	'ZipCode']

data_set = pd.read_csv(filepath,low_memory=True)
data_set.columns = use_cols

data_set.StartTime = pd.to_datetime(data_set.StartTime, format="%Y-%m-%d %H:%M:%S", errors='coerce')
data_set.EndTime = pd.to_datetime(data_set.EndTime, format="%Y-%m-%d %H:%M:%S", errors='coerce')
data_set = data_set.drop(data_set[data_set.StartTime.isnull()].index)
data_set = data_set.drop(data_set[data_set.EndTime.isnull()].index)

print ("Data set is loaded")
print_memory_usage()

airport_mappath="../../AirportToCityState.csv"
airport_df = pd.read_csv(airport_mappath,low_memory=True)
airport_df['State'] = airport_df.CityState.apply(lambda x: x.split('-')[1])
airport_df['City'] = airport_df.CityState.apply(lambda x: x.split('-')[0])
airport_df = airport_df.set_index('Airport')
airport_df = airport_df[['State','City']]
airport_map = airport_df.to_dict('index')

print ("set state-city for weather events")
pool = Pool(cores)
data_set = applyParallel (pool, data_set, State_set,{'map':airport_map.copy()})
pool.close()
pool.join()
print ("setting state-city for weather events done")
print ("*"*80)

invalid_entity = data_set[data_set.City.isnull() | data_set.State.isnull()]
display(invalid_entity.shape)
data_set = data_set.drop(invalid_entity.index)
invalid_entity = data_set[data_set.City.isnull() | data_set.State.isnull()]
display(invalid_entity.shape)
#data_set = data_set.reset_index()

print ("modify start end time type")
pool = Pool(cores)
data_set = parallelize (data_set,type_modif,pool)
pool.close()
pool.join()
print ("type modification for start_end time is done")
print ("*"*80)

print ("compute the duration of events")
pool = Pool(cores)
data_set = parallelize (data_set,days_between,pool)#traffic_events.apply(days_between,axis=1)
pool.close()
pool.join()
#display(data_set)
print_memory_usage()
print ("duration computation is done")

print ("summary of duration:")
summary = data_set.duration.describe()
display(summary)
print ("*"*80)


data_histo = data_set[np.abs(data_set.duration-data_set.duration.mean())<=(10*data_set.duration.std())] 
data_histo.duration/=60
fig, ax = plt.subplots()
data_histo.duration.hist(ax=ax, bins=100, bottom=0.1)
ax.set_yscale('log')
plt.xlabel('event duration in hour')
plt.ylabel('log scale events count')
plt.title('Event Duration Distribution')
plt.savefig('all_event_du_dist.png')

print ("events type:")
d = data_set.groupby('RefinedType').size().to_frame(name='name')
display(d)

#save data_set 
data_set.to_hdf('../../data_set.h5',key='DS_new')
print ("dataset is saved")
