==> long-term original/Process_steps/step4_check_and_remove_conflict.ipynb <==
import argparse
import os
import sys
import psutil
import numpy as np
import pandas as pd
import matplotlib
matplotlib.use('Agg')
import matplotlib.pyplot as plt
#from datetime import datetime
import math
from IPython.display import display
from multiprocessing import cpu_count,Pool 
import multiprocessing
from joblib import Parallel, delayed
from orderedset import OrderedSet
import datetime

long_event = pd.read_hdf('../../long_events_.h5',key='LN_300')

def distance(data,lat,lng,idx):
    fLat = np.radians(data.LocationLat)
    fLon = np.radians(data.LocationLng)
    sLat = np.radians(lat)
    sLon = np.radians(lng)
    R = 3958.7564 #mi
    #R = 6371000.0 #meters
    #R = 6371.0 #km
    
    dLon = sLon - fLon
    dLat = sLat - fLat
    a = np.sin(dLat/2.0)**2 + (np.cos(fLat) * np.cos(sLat) * np.power(np.sin(dLon/2.0), 2))
    
    c = 2.0 * np.arctan2(np.sqrt(a), np.sqrt(1.0 - a))
    #data['dis2event_'+str(idx)] =  
    return R * c
    #return data

def check(data,filepath,key):    
    counter=1
    drop_index=[]
    long_events_ = pd.read_hdf(filepath,key=key)
    for idx,long_event in data.iterrows():
                    collision = data[(long_events_.EventId != long_event.EventId ) & (long_events_.State ==  long_event.State)
                                & 
                                (((long_events_.StartTime >= long_event.StartTime) &
                                    (long_events_.EndTime <= long_event.EndTime))
                                    |
                                    ((long_events_.StartTime < long_event.StartTime) &
                                    (long_events_.EndTime > long_event.StartTime)) 
                                    |
                                    ((long_events_.StartTime < long_event.EndTime) &
                                    (long_events_.EndTime > long_event.EndTime))
                                    |
                                    ((long_events_.StartTime < long_event.StartTime) &
                                    (long_events_.EndTime > long_event.EndTime)))
                                ]

                    if long_event.Type == 'W':
                        collision = collision[collision.AirportCode == long_event.AirportCode]

                    elif long_event.Type != 'W':
                        collison_W = collision[(collision.Type=='W')&(collision.AirportCode == long_event.AirportCode)]
                        collision_T = collision[collision.Type!='W']
                        collision_T = collision_T.assign(distance=pd.Series(
                            distance(collision_T,long_event.LocationLat,long_event.LocationLng,idx)).values)
                        if 'distance' not in collision_T.columns:
                            print ('error')
                        collision_T = collision_T [(collision_T.distance <=raduis)]
                        collision = pd.concat([collison_W,collision_T])
                    else:
                        print ('error: event type ',long_event.Type)
                        raise()
                    if not collision.empty:
                        counter+=1
                        drop_index.append(idx)
                    
    return drop_index

class WithExtraArgs(object):
    def __init__(self, func, **args):
        self.func = func
        self.args = args
    def __call__(self, df):
        return self.func(df, **self.args)
def applyParallel_list(pool,data, func, kwargs):
    data_split = np.array_split(data,partitions)
    data_tag =pool.map(WithExtraArgs(func, **kwargs), data_split)
    return data_tag

def parallelize(data, func,pool):
    data_split = np.array_split(data, partitions)
    df = pool.map(func, data_split)
    return df

raduis = 14 # mi

cores = cpu_count() #Number of CPU cores on your system
partitions = cores #Define as many partitions as you want
print ("number of cores "+str(partitions))

pool = Pool(cores)
drop =  applyParallel_list(pool,long_event,check,{'filepath': '../../long_events_.h5', 'key':'LN_300'})
pool.close()
pool.join()

print (len(drop))

temp = []
for l in drop:
    temp+=l
print (len(temp))
drop_ = temp

long_event_drop1 = long_event.drop(drop_)

print (len(long_event)-len(long_event_drop1))

long_event.loc[drop_[0]]

long_event.loc[drop_[1]]

long_event_drop1.to_hdf('../../long_events_.h5',key='LN_300_nc1')

pool = Pool(cores)
drop = applyParallel_list(pool,long_event_drop1,check,{'filepath': '../../long_events_.h5', 'key':'LN_300_nc1'})
pool.close()
pool.join()

print (drop)

temp2 = []
for l in drop:
    temp2+=l
print (len(temp2))
drop_2 = temp2

con1=long_event_drop1.loc[drop_2[0]]

con2=long_event_drop1.loc[drop_2[1]]

if con2.StartTime < con1.StartTime and con2.EndTime > con1.StartTime:
    print ('yes')

long_event_drop2 = long_event_drop1.drop(drop_2)

print (len(long_event_drop1)-len(long_event_drop2))

long_event_drop2.to_hdf('../../long_events_.h5',key='LN_300_nc2')

pool = Pool(cores)
drop = applyParallel_list(pool,long_event_drop2,check,{'filepath': '../../long_events_.h5', 'key':'LN_300_nc2'})
pool.close()
pool.join()

print (drop)

temp3 = []
for l in drop:
    temp3+=l
print (len(temp3))
drop_3 = temp3

In_LN = pd.read_hdf('../../long_events.h5',key='LN_300')

len(In_LN)

data_set = pd.read_hdf('../../data_set.h5',key='DS_new')

len(data_set)


new_data_set = data_set.drop(In_LN.index)

len(new_data_set)

new_data_set = new_data_set.append(long_event_drop2)

len(new_data_set)

new_data_set.to_hdf('../../data_set_.h5',key='DS_300')

long_event_drop2.duration.max()

52435.01666666667/60/24

long_event_drop2.duration.min()

long_event_drop2.duration.mean()

new_data_set.index.get_duplicates()
