==> long-term original/Process_steps/step7_merge_dict_2_list.ipynb <==
import argparse
import os
import sys
import psutil
import numpy as np
import pandas as pd
import matplotlib
matplotlib.use('Agg')
import matplotlib.pyplot as plt
import math
from IPython.display import display
from multiprocessing import cpu_count,Pool 
import multiprocessing
from joblib import Parallel, delayed
from orderedset import OrderedSet
import datetime
import pickle
import glob

path = "dict_files/"
dirpath = os.getcwd()
dirpath+='/'+path
length=0
print (dirpath)
final_dict={}
for filename in glob.glob(os.path.join(dirpath, '*.pkl')):
    print (filename)
    temp_dict = pickle.load( open( filename, "rb" ))
    length+=len(temp_dict)
    if temp_dict.keys() & final_dict.keys():
        print ('\t\t conflict')
        #print (temp_dict.keys() & final_dict.keys())
    else:
        final_dict.update(temp_dict)


if len(final_dict) != length:
    print ('error')

print (len(final_dict))
print (length)


long_events_ = pd.read_hdf('../../long_events_.h5',key='LN_300_nc2')

if long_events_.shape[0] != len(final_dict):
    print ('error')
print (len(long_events_))

final_list=[]
for idx,long_event in long_events_.iterrows():
    final_list.append(final_dict[idx])
print (len(final_list))

def splitter(A):
    B = A[0:len(A)//2]
    C = A[len(A)//2:]

    return (B,C)

half1,half2 = splitter(final_list)

len(half1)+len(half2)

np.save("../../event_list_45_300_1",half1)

np.save("../../event_list_45_300_2",half2)

arr = np.asarray(final_list)

arr

final_dict
f = open('../../final_dict.pkl',"wb")
pickle.dump(final_dict,f)
f.close()
